Finished all User functionality of app, have to finish admin stuff

+++ src/main/java/com/revature/ers/dao/ReimbursementDaoImp.java
@@ -1,8 +1,6 @@
 package com.revature.ers.dao;
 
 import com.revature.ers.models.Reimbursement;
-import com.revature.ers.models.Status;
-import com.revature.ers.models.Type;
 import com.revature.ers.models.User;
 import com.revature.ers.util.ConnectionManager;
 
@@ -12,6 +10,7 @@ import java.util.*;
 
 
 public class ReimbursementDaoImp implements ReimbursementDao{
+
 
 
     @Override
@@ -39,20 +38,19 @@ public class ReimbursementDaoImp implements ReimbursementDao{
 
     @Override
     public List<Reimbursement> getReimbursementUser(User u) {
-
         List<Reimbursement> list = new ArrayList<>();
 
         try{
             String sql = "SELECT er.REIMB_ID, er.REIMB_AMOUNT, er.REIMB_SUBMITTED, er.REIMB_RESOLVED, eu.username as resolver_username  , ers_reimbursement_status.reimb_status, ers_reimbursement_type.reimb_type "
-                + "from ers_reimbursement er "
-                + "join ers_reimbursement_status "
-                + "on er.reimb_status_id = ers_reimbursement_status.reimb_status_id "
-                + "join ers_reimbursement_type "
-                + "on er.reimb_type_id = ers_reimbursement_type.reimb_type_id "
-                + "left join ers_users eu "
-                + "on er.reimb_resolver = eu.user_id "
-                + "where er.reimb_author = ? "
-                + "order by er.reimb_id ";
+                    + "from ers_reimbursement er "
+                    + "join ers_reimbursement_status "
+                    + "on er.reimb_status_id = ers_reimbursement_status.reimb_status_id "
+                    + "join ers_reimbursement_type "
+                    + "on er.reimb_type_id = ers_reimbursement_type.reimb_type_id "
+                    + "left join ers_users eu "
+                    + "on er.reimb_resolver = eu.user_id "
+                    + "where er.reimb_author = ? "
+                    + "order by er.reimb_id ";
             PreparedStatement pstmt = ConnectionManager.getConnection().prepareStatement(sql);
             pstmt.setInt(1, u.getId());
 
@@ -63,7 +61,12 @@ public class ReimbursementDaoImp implements ReimbursementDao{
                 r.setAmount(results.getInt("reimb_amount"));
                 r.setSubmitted(results.getDate("reimb_submitted"));
                 r.setResolved(results.getDate("reimb_resolved"));
-                //r.setResolver(results.getString("resolver_username"));
+                r.setAuthor(u);
+
+
+                //have to set up a way to set a resolver on local when receiving from a remote database
+
+
                 r.setStatus(results.getString("reimb_status"));
                 r.setType(results.getString("reimb_type"));
                 list.add(r);
@@ -113,7 +116,83 @@ public class ReimbursementDaoImp implements ReimbursementDao{
     }
 
     @Override
-    public void updateReimbursement(Reimbursement r, User u) {
+    public List<Reimbursement> getReimbursementByStatus(int status) {
+        List<Reimbursement> list = new ArrayList<>();
+        String sql = "SELECT er.REIMB_ID, er.REIMB_AMOUNT, er.REIMB_SUBMITTED, er.REIMB_RESOLVED, a.username as author_username, r.username as resolver_username  , ers_reimbursement_status.reimb_status, ers_reimbursement_type.reimb_type "
+                + "from ers_reimbursement er "
+                + "join ers_reimbursement_status "
+                + "on er.reimb_status_id = ers_reimbursement_status.reimb_status_id "
+                + "join ers_reimbursement_type"
+                + "on er.reimb_type_id = ers_reimbursement_type.reimb_type_id "
+                + "left join ers_users a "
+                + "on er.reimb_AUTHOR = a.user_id "
+                + "left join ers_users r "
+                + "on er.reimb_resolver = r.user_id "
+                + "where er.reimb_status_id = ? ";
+        try{
+            PreparedStatement pstmt = ConnectionManager.getConnection().prepareStatement(sql);
+            pstmt.setInt(1, status);
+
+            ResultSet results = pstmt.executeQuery();
+
+            while(results.next()){
+                Reimbursement r = new Reimbursement();
+                r.setId(results.getInt("reimb_id"));
+                r.setAmount(results.getInt("reimb_amount"));
+                r.setSubmitted(results.getDate("reimb_submitted"));
+                r.setResolved(results.getDate("reimb_resolved"));
+                //r.setAuthor(results.getString("author_username"));
+                //r.setResolver(results.getString("resolver_username"));
+                r.setStatus(results.getString("reimb_status"));
+                r.setType(results.getString("reimb_type"));
+                list.add(r);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return list;
+    }
+
+    @Override
+    public List<Reimbursement> getPendingReimbursementByUser(User u) {
+        int id = u.getId();
+        List<Reimbursement> list = new ArrayList<>();
+        String sql = "SELECT er.REIMB_ID, er.REIMB_AMOUNT, er.REIMB_SUBMITTED, er.REIMB_RESOLVED, a.username as author_username, r.username as resolver_username  , ers_reimbursement_status.reimb_status, ers_reimbursement_type.reimb_type"
+                + " from ers_reimbursement er"
+                + " join ers_reimbursement_status"
+                + " on er.reimb_status_id = ers_reimbursement_status.reimb_status_id"
+                + " join ers_reimbursement_type"
+                + " on er.reimb_type_id = ers_reimbursement_type.reimb_type_id"
+                + " left join ers_users a"
+                + " on er.reimb_AUTHOR = a.user_id"
+                + " left join ers_users r"
+                + " on er.reimb_resolver = r.user_id"
+                + " where er.reimb_status_id  = 0 and er.reimb_author = ?";
+        try{
+            PreparedStatement pstmt = ConnectionManager.getConnection().prepareStatement(sql);
+            pstmt.setInt(1, id);
+
+            ResultSet results = pstmt.executeQuery();
+
+            while(results.next()){
+                Reimbursement r = new Reimbursement();
+                r.setId(results.getInt("reimb_id"));
+                r.setAmount(results.getInt("reimb_amount"));
+                r.setSubmitted(results.getDate("reimb_submitted"));
+                r.setResolved(results.getDate("reimb_resolved"));
+                r.setAuthor(u);
+                r.setStatus(results.getString("reimb_status"));
+                r.setType(results.getString("reimb_type"));
+                list.add(r);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return list;
+    }
+
+    @Override
+    public void completeReimbursement(Reimbursement r, User u) {
         String sql = "UPDATE ERS_REIMBURSEMENT SET reimb_status_id=?, reimb_resolver=?  WHERE reimb_id = ?";
         try {
             PreparedStatement pstmt = ConnectionManager.getConnection().prepareStatement(sql);
@@ -127,12 +206,27 @@ public class ReimbursementDaoImp implements ReimbursementDao{
     }
 
     @Override
-    public void cancelReimbursement(int id) {
-        String sql = "DELETE FROM ERS_REIMBURSEMENT WHERE REIMB_ID=?";
+    public void editReimbursement(double amount, User u, int id){
+        String sql = "UPDATE ERS_REIMBURSEMENT SET reimb_amount=? WHERE reimb_author=? and reimb_id=?";
+        try{
+            PreparedStatement pstmt = ConnectionManager.getConnection().prepareStatement(sql);
+            pstmt.setDouble(1, amount);
+            pstmt.setInt(2, u.getId());
+            pstmt.setInt(3, id);
+            pstmt.executeUpdate();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Override
+    public void cancelReimbursement(int id, User u) {
+        String sql = "DELETE FROM ERS_REIMBURSEMENT WHERE REIMB_ID=? AND reimb_author=?";
 
         try{
             PreparedStatement pstmt = ConnectionManager.getConnection().prepareStatement(sql);
             pstmt.setInt(1, id);
+            pstmt.setInt(2, u.getId());
             pstmt.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
